fix warnings

+++ cpp/OpenGL/Utils/Grid2D.cpp
@@ -386,7 +386,7 @@ static const float INV = std::numeric_limits<float>::max();
 static const Vec2ui NO_POS{std::numeric_limits<uint32_t>::max(),
                     std::numeric_limits<uint32_t>::max()};
 static float dist(size_t x, size_t y, const Vec2ui& p) {
-  return std::sqrt( (x-p.x())*(x-p.x()) + (y-p.y())*(y-p.y()) );
+  return std::sqrtf( (x-p.x())*(x-p.x()) + (y-p.y())*(y-p.y()) );
 }
 
 Grid2D Grid2D::toSignedDistance(float threshold) const {
@@ -470,6 +470,6 @@ Grid2D Grid2D::toSignedDistance(float threshold) const {
 
 GLTexture2D Grid2D::toTexture() const {
   GLTexture2D result;
-  result.setData(data, width, height, 1);
+  result.setData(data, uint32_t(width), uint32_t(height), 1);
   return result;
 }
